move run_deployment
